+++ src/components/strategy/index.ts
@@ -1,0 +1,2 @@
+export * from './CreateStrategy.tsx';
+export * from './News.tsx';