+++ src/models/index.ts
@@ -3,3 +3,4 @@ export * from './user.model.ts';
 export * from './error.model.ts';
 export * from './post.model.ts';
 export * from './zustand.model.ts';
+export * from './strategy.model.ts';