+++ CHANGELOG.md
@@ -2,6 +2,23 @@
 
 This document lists breaking changes for each major release.
 
+## v.0.0.6 (18/12/2024)
+
+Feature, KAN-3, KAN-50, KAN-44: Create Strategy, profile, pagination.
+
+- Create adaptors to strategy.
+- Create Data range componnent to look from to post order.
+- Position down the config.
+- Create component CreateStrategy to handle the strategy.
+- Pagination in Posts component and hookd to handle logic.
+- Create interface to strategy, posts.
+- Create Profile component.
+- Create news skeleton.
+- Add private route /account.
+- Update services to accept query params.
+- Create validator strategy.
+- Create initial values about Strategy.
+
 ## v.0.0.5 (10/12/2024)
 
 Feature, KAN-35, KAN-47: Add language, dark mode selector.